+++ cloud-controller-manager/do/loadbalancers_test.go
@@ -6220,6 +6220,54 @@ func Test_buildFirewall(t *testing.T) {
 				Allow: []string{"ip:1.2.3.4", "ip:1.2.3.5"},
 			},
 		},
+		{
+			name: "source ranges not set",
+			service: &v1.Service{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "test",
+					UID:  "abc123",
+				},
+				Spec: v1.ServiceSpec{
+					LoadBalancerSourceRanges: []string{},
+				},
+			},
+			expectedFirewall: &godo.LBFirewall{},
+		},
+		{
+			name: "source ranges set",
+			service: &v1.Service{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "test",
+					UID:  "abc123",
+				},
+				Spec: v1.ServiceSpec{
+					LoadBalancerSourceRanges: []string{"1.2.0.0/16"},
+				},
+			},
+			expectedFirewall: &godo.LBFirewall{
+				Allow: []string{"cidr:1.2.0.0/16"},
+			},
+		},
+		{
+			name: "source ranges and annotations set",
+			service: &v1.Service{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "test",
+					UID:  "abc123",
+					Annotations: map[string]string{
+						annDODenyRules:  "cidr:1.2.0.0/16",
+						annDOAllowRules: "ip:1.2.3.4,ip:1.2.3.5",
+					},
+				},
+				Spec: v1.ServiceSpec{
+					LoadBalancerSourceRanges: []string{"1.3.0.0/16"},
+				},
+			},
+			expectedFirewall: &godo.LBFirewall{
+				Deny:  []string{"cidr:1.2.0.0/16"},
+				Allow: []string{"cidr:1.3.0.0/16"},
+			},
+		},
 	}
 
 	for _, test := range testcases {