+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+* Adding support for specifiying `loadBalancerSourceRanges` in the service spec. Source ranges take precedence over annotation based allow rules (`service.beta.kubernetes.io/do-loadbalancer-allow-rules`).
+
 ## v0.1.51 (beta) - May 28, 2024
 
 * Adjusts load balancer health check behaviour to probe Kubernetes components correctly, ensuring that LB traffic stops